Update README.md
This project is an emulator of an online shop that sells programming courses. It focuses only on the backend side, and it was a group project for a university subject called Methodology of  Programming II. The project was a collaboration between me and 2 of my classmates, Dita Pelaj and Egor Kolonyuk under the supervision of Professor Zahra Ahmadipour.